requirements.txt
Flask
Flask-Mail
Flask-WTF
python-dotenv
email_validator